build.tcl: Fix OOC synthesis error of block designs
In Vivado 2017.4 the get_property command fails, because get_files does
not return the .bd file although the file was already listed in the
get_files command of the foreach loop. The second get_files still does
not make any sense and using the bd variable of the foreach loop does work.

Signed-off-by: Andreas Braun <andreas.braun@missinglinkelectronics.com>