Document synthesis and implementation options
The configuration dictionary keys SYNTH_STRAT and IMPL_STRAT have not
been documented, because the user should have used direct TCL commands
with the exec_tcl option or by using a hook. But changing the synthesis
or implementation strategies does also remove the build system hooks
from the run so that the build fails. By applying the strategies before
the user hooks are set or exec_tcl is being executed (which is done by
the SYNTH_STRAT and IMPL_STRAT keys) the hooks are set correctly.

Signed-off-by: Andreas Braun <andreas.braun@missinglinkelectronics.com>