build.tcl: Remove exit(0), modify end_step instead
The build targets 'package', 'sim' and 'sim_prep' terminate successfully with
exit(0). Therefore, the Vivado log parsing for errors and warnings is skipped.
Instead of exiting, the build end step is now set to the targeted step so that
all following build steps are skipped and the report parsing is done.

Signed-off-by: Andreas Braun <andreas.braun@missinglinkelectronics.com>