hooks/: Fix wrong time stamp in name of build results
The time stamp of build result names is not correct when a build is started with a start step other than 'prj'. The build system hooks need to update the build time stamp only when the start step matches the hook step. Due to a fixed value for the build start step encoding, the references have become wrong as steps were inserted in the past.

Signed-off-by: Andreas Braun <andreas.braun@missinglinkelectronics.com>